requirements.txt
streamlit==1.31.0
pandas==2.1.0
openai==1.3.0
trafilatura==1.6.0
requests==2.31.0
beautifulsoup4==4.12.0
dnspython==2.4.0
openpyxl==3.1.2
xlsxwriter==3.1.0
